requirements.txt
requests
textblob
python-dotenv

